+++ python/gtirb/auxdata.py
@@ -1,11 +1,44 @@
-import typing
 from io import BytesIO
+from typing import Any, ClassVar, Dict, Optional
 from uuid import UUID
 
 from .node import Node
 from .proto import AuxData_pb2
 from .serialization import Serialization
 from .util import DictLike
+
+
+class _LazyDataContainer:
+    """
+    Container that holds the raw byte stream until it is read, then releases
+    it. If it is never read, then serialization skips re-encoding (and
+    deserializing) the data.
+    """
+
+    def __init__(self, raw_data, type_name, get_by_uuid):
+        self.raw_data = raw_data
+        self.type_name = type_name
+        self.get_by_uuid = get_by_uuid
+
+    def get_data(self):
+        # type: (Any) -> Any
+        """
+        Get any pending still-serialized data, or return the passed data
+        instead (the default).
+        """
+        assert self.raw_data is not None
+        rv = AuxData.serializer.decode(
+            self.raw_data, self.type_name, self.get_by_uuid
+        )
+        self.raw_data = None
+        return rv
+
+    def get_raw_data(self):
+        # type: () -> bytes
+        """
+        """
+        assert self.raw_data is not None
+        return self.raw_data
 
 
 class AuxData:
@@ -26,45 +59,72 @@ class AuxData:
         Used to determine the proper codec for serializing this AuxData.
     """
 
-    serializer = Serialization()  # type: typing.ClassVar[Serialization]
+    serializer = Serialization()  # type: ClassVar[Serialization]
     """This is a :class:`gtirb.Serialization` instance, used to
     encode and decode ``data`` fields of all ``AuxData``. See
     :mod:`gtirb.serialization` for details.
     """
 
-    def __init__(self, data, type_name):
-        # type: (typing.Any, str) -> None
+    def __init__(self, data, type_name, lazy_container=None):
+        # type: (Any, str, Optional[_LazyDataContainer]) -> None
         """
         :param data: The value stored in this AuxData.
         :param type_name: A string describing the type of ``data``.
             Used to determine the proper codec for serializing this AuxData.
+        :param lazy_container: An object that will lazily deserialize the
+            auxdata table backing this object, or None.
         """
-
-        self.data = data  # type: typing.Any
+        self._lazy_container = lazy_container
+        self._data = data  # type: Any
         self.type_name = type_name  # type: str
+
+    @property
+    def data(self):
+        if self._lazy_container is not None:
+            self._data = self._lazy_container.get_data()
+            self._lazy_container = None
+        return self._data
+
+    @data.setter
+    def data(self, value):
+        self._data = value
+        self._lazy_container = None
 
     @classmethod
     def _from_protobuf(cls, aux_data, ir):
-        # type: (AuxData_pb2.AuxData, typing.Optional["IR"]) -> AuxData
-        """Deserialize AuxData from Protobuf.
+        # type: (AuxData_pb2.AuxData, Optional["IR"]) -> AuxData
+        """Deserialize AuxData from Protobuf. Lazy, will not perform
+        deserialization until .data is accessed.
 
         :param aux_data: The Protobuf AuxData object.
         """
 
-        data = AuxData.serializer.decode(
-            BytesIO(aux_data.data), aux_data.type_name, ir.get_by_uuid
+        # Defer deserialization until someone accesses .data
+        lazy_container = _LazyDataContainer(
+            aux_data.data, aux_data.type_name, ir.get_by_uuid
         )
-        return cls(data=data, type_name=aux_data.type_name)
+        return cls(
+            data=None,
+            type_name=aux_data.type_name,
+            lazy_container=lazy_container,
+        )
 
     def _to_protobuf(self):
         # type: () -> AuxData_pb2.AuxData
         """Get a Protobuf representation of the AuxData."""
 
-        out_bytes_array = BytesIO()
-        AuxData.serializer.encode(out_bytes_array, self.data, self.type_name)
         proto_auxdata = AuxData_pb2.AuxData()
         proto_auxdata.type_name = self.type_name
-        proto_auxdata.data = out_bytes_array.getvalue()
+        # If we are serializing the same data, and the way that data is encoded
+        # has not changed, then just use the already serialized copy.
+        if self._lazy_container is not None and (
+            self.type_name == self._lazy_container.type_name
+        ):
+            proto_auxdata.data = self._lazy_container.get_raw_data()
+        else:
+            data_stream = BytesIO()
+            AuxData.serializer.encode(data_stream, self.data, self.type_name)
+            proto_auxdata.data = data_stream.getvalue()
         return proto_auxdata
 
     def __repr__(self):
@@ -73,7 +133,7 @@ class AuxData:
             "AuxData("
             "type_name={type_name!r}, "
             "data={data!r}, "
-            ")".format(**self.__dict__)
+            ")".format(type_name=self.type_name, data=self.data)
         )
 
 
@@ -89,7 +149,7 @@ class AuxDataContainer(Node):
     def __init__(
         self,
         aux_data={},  # type: DictLike[str, AuxData]
-        uuid=None,  # type: typing.Optional[UUID]
+        uuid=None,  # type: Optional[UUID]
     ):
         # type: (...) -> None
         """
@@ -101,11 +161,11 @@ class AuxDataContainer(Node):
             Defaults to None.
         """
         super().__init__(uuid)
-        self.aux_data = dict(aux_data)  # type: typing.Dict[str, AuxData]
+        self.aux_data = dict(aux_data)  # type: Dict[str, AuxData]
 
     @classmethod
     def _read_protobuf_aux_data(cls, proto_container, ir):
-        # type: (typing.Any,typing.Optional["IR"]) -> typing.Dict[str, AuxData]
+        # type: (Any,Optional["IR"]) -> Dict[str, AuxData]
         """
         Instead of the overrided _decode_protobuf, this method requires the
         Protobuf message to read from. AuxDataContainers need to call this
@@ -120,7 +180,7 @@ class AuxDataContainer(Node):
         }
 
     def _write_protobuf_aux_data(self, proto_container):
-        # type: (typing.Any) -> None
+        # type: (Any) -> None
         """
         Instead of the overrided _to_protobuf, this method requires the
         Protobuf message to write into. AuxDataContainers need to call this
@@ -133,7 +193,7 @@ class AuxDataContainer(Node):
             proto_container.aux_data[k].CopyFrom(v._to_protobuf())
 
     def deep_eq(self, other):
-        # type: (typing.Any) -> bool
+        # type: (Any) -> bool
         """This overrides :func:`gtirb.Node.deep_eq` to check for
         AuxData equality.